--- a/ingestr/src/filters.py
+++ b/ingestr/src/filters.py
@@ -30,7 +30,7 @@ def handle_mysql_empty_dates(row):
 
 def table_adapter_exclude_columns(cols: list[str]):
     from dlt.common.libs.sql_alchemy import Table
-    
+
     def excluder(table: Table):
         cols_to_remove = [col for col in table._columns if col.name in cols]  # type: ignore
         for col in cols_to_remove: