--- a/ingestr/src/partition.py
+++ b/ingestr/src/partition.py
@@ -1,7 +1,6 @@
 from typing import Dict
 
 from dlt.common.schema.typing import TColumnSchema
-
 from dlt.sources import DltResource, DltSource
 
 import ingestr.src.resource as resource
@@ -13,7 +12,7 @@ def apply_athena_hints(
     additional_hints: Dict[str, TColumnSchema] = {},
 ) -> None:
     from dlt.destinations.adapters import athena_adapter, athena_partition
-    
+
     def _apply_partition_hint(resource: DltResource) -> None:
         columns = resource.columns if resource.columns else {}