This PR is the last step of making all the imports lazy.

Before:
```
  Time (mean ± σ):      1.803 s ±  0.108 s    [User: 1.586 s, System: 0.192 s]
  Range (min … max):    1.716 s …  2.216 s    50 runs
```

After all the lazy imports:
```
  Time (mean ± σ):     787.6 ms ± 210.6 ms    [User: 586.0 ms, System: 188.7 ms]
  Range (min … max):   662.6 ms … 1373.2 ms    50 runs
```

The difference is ~55% speed up in barebones import timings. I don't think the difference would be this significant in actual execution, but the commands already started feeling a lot more snappy.